requirements.txt
absl-py==0.13.0
aiodns==3.0.0
aiofiles==0.7.0
aiogrpc==1.8
aiohttp==3.7.4.post0
aiohttp-proxy==0.1.2
aiohttp-socks==0.6.0
aioredis==2.0.0
aiosqlite==0.17.0
astroid==2.6.5
astunparse==1.6.3
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.7
beautifulsoup4==4.9.3
brotlipy==0.7.0
cachetools==4.2.2
cchardet==2.1.7
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
discord.py==1.7.3
flatbuffers==1.12
gast==0.4.0
google-auth==1.32.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
idna==2.10
isort==5.9.2
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
lazy-object-proxy==1.6.0
lxml==4.6.3
Markdown==3.3.4
mccabe==0.6.1
multidict==5.1.0
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
Pillow==8.3.1
PixivPy-Async==1.2.12
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==4.0.0
pycodestyle==2.7.0
pycparser==2.20
pylint==2.9.6
PyNaCl==1.4.0
pysaucenao==1.6.1
python-socks==1.2.4
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
soupsieve==2.2.1
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
tensorflow-hub==0.12.0
termcolor==1.1.0
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
yarl==1.6.3
youtube-dl==2021.6.6
